examples/plain-agents/requirements.txt
# OpenTelemetry Python SDK and OTLP exporters
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-grpc>=1.20.0

# Required for instrumentation
opentelemetry-instrumentation>=0.41b0

# HTTP client for telemetry validation
requests>=2.31.0

# YAML configuration parsing
PyYAML>=6.0

# Disable SSL warnings for development
urllib3>=2.0.0

# Testing dependencies
pytest>=7.4.0
hypothesis>=6.82.0
